Added RenderManager, bug fixes, start of refactoring all code to ECS instead

+++ Core/src/Time.cpp
@@ -1,8 +1,8 @@
 #include "Time.h"
 
-float Time::_deltaTime = 0;
+double Time::_deltaTime = 0;
 
-void Time::UpdateDeltaTime(float currentTime)
+void Time::UpdateDeltaTime(double currentTime)
 {
     static float lastFrame = 0.0f;
     float currentFrame = currentTime;
